pyproject.toml
[project]
name = "maldoc"
version = "0.1.0"
description = "Adversarial document generation and evaluation framework for AI pipeline security testing"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.15",
    "pydantic>=2.0",
    "httpx>=0.28",
    "fpdf2>=2.8",
    "python-docx>=1.1",
    "Pillow>=11.0",
    "rich>=13.0",
]

[project.scripts]
maldoc = "maldoc.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/maldoc"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires Docker and Ollama",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-httpx>=0.35",
]
